+++ crates/engine/src/lib.rs
@@ -1,6 +1,7 @@
 #![warn(clippy::all)]
 #![warn(clippy::pedantic)]
 
+mod builder;
 mod engine;
 mod error;
 mod plugin;